requirements.txt
google-generativeai=0.8.4
python-dotenv=1.0.1
requests=2.32.3
google-api-python-client=2.161.0